Add files via upload
Coin functionality works, price of food items are presented on the buttons. CSS added to readme file, using "ul li" element for the lists, looked it up on MDN and it works fine. Some balancing to the numbers. Next is to try and get different images show depending on the user input and change the layout to work with that.